+++ data/data_analysis.md
@@ -1,4 +1,3 @@
-
 # Retinal Disease Classification – Data Analysis
 
 This document contains extended analysis and visualizations that support the Updates.md in `Updates.md`.
@@ -70,4 +69,3 @@ img_transforms = {
 ```
 
 ---
-
